Fixed randPoint

+++ src/main/kotlin/Main.kt
@@ -64,7 +64,7 @@ class Point(val x: Double, val y: Double, var cluster: Int = -1){
 }
 
 fun randPoint(x1:Double, x2:Double, y1:Double, y2:Double):() -> Point{
-    return({-> Point((x1 +(x2-x1) * Random.nextFloat()).toDouble(), (y1 + (y2-y1) * Random.nextFloat()).toDouble())})
+    return({-> Point((x1 +(x2-x1) * Random.nextDouble()), (y1 + (y2-y1) * Random.nextDouble()))})
 }
 
 // Problem 3
